+++ release.sh
@@ -1,0 +1,19 @@
+#!/bin/bash
+
+set -e
+echo "Enter release version: "
+read VERSION
+
+read -p "Releasing $VERSION - are you sure (y/n)" -n 1 -r
+echo #
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  echo "Releasing $VERSION ..."
+
+  VERSION=$VERSION
+  git add -A
+  git commit -m "[build] $VERSION"
+  npm version $VERSION --message "[release] $VERSION"
+
+  git push
+fi